Split the Python code from __init__.py into submodules.
Try to keep __init__.py still compatible with old rarfile.py
by re-exporting old symbols.
