+++ src/rarfile/bits.py
@@ -1,0 +1,178 @@
+"""RAR constants
+"""
+
+# block types
+RAR_BLOCK_MARK = 0x72           # r
+RAR_BLOCK_MAIN = 0x73           # s
+RAR_BLOCK_FILE = 0x74           # t
+RAR_BLOCK_OLD_COMMENT = 0x75    # u
+RAR_BLOCK_OLD_EXTRA = 0x76      # v
+RAR_BLOCK_OLD_SUB = 0x77        # w
+RAR_BLOCK_OLD_RECOVERY = 0x78   # x
+RAR_BLOCK_OLD_AUTH = 0x79       # y
+RAR_BLOCK_SUB = 0x7a            # z
+RAR_BLOCK_ENDARC = 0x7b         # {
+
+# flags for RAR_BLOCK_MAIN
+RAR_MAIN_VOLUME = 0x0001
+RAR_MAIN_COMMENT = 0x0002
+RAR_MAIN_LOCK = 0x0004
+RAR_MAIN_SOLID = 0x0008
+RAR_MAIN_NEWNUMBERING = 0x0010
+RAR_MAIN_AUTH = 0x0020
+RAR_MAIN_RECOVERY = 0x0040
+RAR_MAIN_PASSWORD = 0x0080
+RAR_MAIN_FIRSTVOLUME = 0x0100
+RAR_MAIN_ENCRYPTVER = 0x0200
+
+# flags for RAR_BLOCK_FILE
+RAR_FILE_SPLIT_BEFORE = 0x0001
+RAR_FILE_SPLIT_AFTER = 0x0002
+RAR_FILE_PASSWORD = 0x0004
+RAR_FILE_COMMENT = 0x0008
+RAR_FILE_SOLID = 0x0010
+RAR_FILE_DICTMASK = 0x00e0
+RAR_FILE_DICT64 = 0x0000
+RAR_FILE_DICT128 = 0x0020
+RAR_FILE_DICT256 = 0x0040
+RAR_FILE_DICT512 = 0x0060
+RAR_FILE_DICT1024 = 0x0080
+RAR_FILE_DICT2048 = 0x00a0
+RAR_FILE_DICT4096 = 0x00c0
+RAR_FILE_DIRECTORY = 0x00e0
+RAR_FILE_LARGE = 0x0100
+RAR_FILE_UNICODE = 0x0200
+RAR_FILE_SALT = 0x0400
+RAR_FILE_VERSION = 0x0800
+RAR_FILE_EXTTIME = 0x1000
+RAR_FILE_EXTFLAGS = 0x2000
+
+# flags for RAR_BLOCK_ENDARC
+RAR_ENDARC_NEXT_VOLUME = 0x0001
+RAR_ENDARC_DATACRC = 0x0002
+RAR_ENDARC_REVSPACE = 0x0004
+RAR_ENDARC_VOLNR = 0x0008
+
+# flags common to all blocks
+RAR_SKIP_IF_UNKNOWN = 0x4000
+RAR_LONG_BLOCK = 0x8000
+
+# Subtypes for RAR_BLOCK_OLD_SUB
+RAR_OLD_SUB_OS2 = 0x100
+RAR_OLD_SUB_UNIX = 0x101
+RAR_OLD_SUB_MAC = 0x102
+RAR_OLD_SUB_BEOS = 0x103
+RAR_OLD_SUB_NT = 0x104
+RAR_OLD_SUB_STREAM = 0x105
+
+# Host OS types
+RAR_OS_MSDOS = 0    #: MSDOS (only in RAR3)
+RAR_OS_OS2 = 1      #: OS2 (only in RAR3)
+RAR_OS_WIN32 = 2    #: Windows
+RAR_OS_UNIX = 3     #: UNIX
+RAR_OS_MACOS = 4    #: MacOS (only in RAR3)
+RAR_OS_BEOS = 5     #: BeOS (only in RAR3)
+
+# Compression methods - "0".."5"
+RAR_M0 = 0x30   #: No compression.
+RAR_M1 = 0x31   #: Compression level `-m1` - Fastest compression.
+RAR_M2 = 0x32   #: Compression level `-m2`.
+RAR_M3 = 0x33   #: Compression level `-m3`.
+RAR_M4 = 0x34   #: Compression level `-m4`.
+RAR_M5 = 0x35   #: Compression level `-m5` - Maximum compression.
+
+RAR_MAX_PASSWORD = 127  #: Max number of utf-16 chars in passwords.
+RAR_MAX_KDF_SHIFT = 24  #: Max power-of-2 for KDF count
+
+#
+# RAR5 constants
+#
+
+RAR5_BLOCK_MAIN = 1
+RAR5_BLOCK_FILE = 2
+RAR5_BLOCK_SERVICE = 3
+RAR5_BLOCK_ENCRYPTION = 4
+RAR5_BLOCK_ENDARC = 5
+
+RAR5_BLOCK_FLAG_EXTRA_DATA = 0x01
+RAR5_BLOCK_FLAG_DATA_AREA = 0x02
+RAR5_BLOCK_FLAG_SKIP_IF_UNKNOWN = 0x04
+RAR5_BLOCK_FLAG_SPLIT_BEFORE = 0x08
+RAR5_BLOCK_FLAG_SPLIT_AFTER = 0x10
+RAR5_BLOCK_FLAG_DEPENDS_PREV = 0x20
+RAR5_BLOCK_FLAG_KEEP_WITH_PARENT = 0x40
+
+RAR5_MAIN_FLAG_ISVOL = 0x01
+RAR5_MAIN_FLAG_HAS_VOLNR = 0x02
+RAR5_MAIN_FLAG_SOLID = 0x04
+RAR5_MAIN_FLAG_RECOVERY = 0x08
+RAR5_MAIN_FLAG_LOCKED = 0x10
+
+RAR5_FILE_FLAG_ISDIR = 0x01
+RAR5_FILE_FLAG_HAS_MTIME = 0x02
+RAR5_FILE_FLAG_HAS_CRC32 = 0x04
+RAR5_FILE_FLAG_UNKNOWN_SIZE = 0x08
+
+RAR5_COMPR_SOLID = 0x40
+
+RAR5_ENC_FLAG_HAS_CHECKVAL = 0x01
+
+RAR5_ENDARC_FLAG_NEXT_VOL = 0x01
+
+RAR5_XFILE_ENCRYPTION = 1
+RAR5_XFILE_HASH = 2
+RAR5_XFILE_TIME = 3
+RAR5_XFILE_VERSION = 4
+RAR5_XFILE_REDIR = 5
+RAR5_XFILE_OWNER = 6
+RAR5_XFILE_SERVICE = 7
+
+RAR5_XTIME_UNIXTIME = 0x01
+RAR5_XTIME_HAS_MTIME = 0x02
+RAR5_XTIME_HAS_CTIME = 0x04
+RAR5_XTIME_HAS_ATIME = 0x08
+RAR5_XTIME_UNIXTIME_NS = 0x10
+
+RAR5_XENC_CIPHER_AES256 = 0
+
+RAR5_XENC_CHECKVAL = 0x01
+RAR5_XENC_TWEAKED = 0x02
+
+RAR5_XHASH_BLAKE2SP = 0
+
+RAR5_XREDIR_UNIX_SYMLINK = 1
+RAR5_XREDIR_WINDOWS_SYMLINK = 2
+RAR5_XREDIR_WINDOWS_JUNCTION = 3
+RAR5_XREDIR_HARD_LINK = 4
+RAR5_XREDIR_FILE_COPY = 5
+
+RAR5_XREDIR_ISDIR = 0x01
+
+RAR5_XOWNER_UNAME = 0x01
+RAR5_XOWNER_GNAME = 0x02
+RAR5_XOWNER_UID = 0x04
+RAR5_XOWNER_GID = 0x08
+
+RAR5_OS_WINDOWS = 0
+RAR5_OS_UNIX = 1
+
+DOS_MODE_ARCHIVE = 0x20
+DOS_MODE_DIR = 0x10
+DOS_MODE_SYSTEM = 0x04
+DOS_MODE_HIDDEN = 0x02
+DOS_MODE_READONLY = 0x01
+
+RAR5_PW_CHECK_SIZE = 8
+RAR5_PW_SUM_SIZE = 4
+
+##
+## internal constants
+##
+
+RAR_ID = b"Rar!\x1a\x07\x00"
+RAR5_ID = b"Rar!\x1a\x07\x01\x00"
+
+RAR_V3 = 3
+RAR_V5 = 5
+
+__all__ = tuple(n for n in globals() if n.startswith(("RAR", "DOS")))
