+++ src/rarfile/errors.py
@@ -1,3 +1,6 @@
+"""Custom error classes.
+"""
+
 __all__ = (
     "Error",
     "BadRarFile",
