+++ test/test_crypto.py
@@ -1,13 +1,11 @@
 """Crypto tests.
 """
 
-from __future__ import division, print_function
-
 from binascii import unhexlify
 
 import pytest
 
-import rarfile
+from rarfile.crypto import AES_CBC_Decrypt, have_crypto
 
 try:
     from cryptography.hazmat.backends import default_backend
@@ -23,7 +21,7 @@ except ImportError:
     pass
 
 
-@pytest.mark.skipif(not rarfile._have_crypto, reason="No crypto")
+@pytest.mark.skipif(not have_crypto, reason="No crypto")
 def test_aes128_cbc():
     data = b"0123456789abcdef" * 2
     key = b"\x02" * 16
@@ -32,11 +30,11 @@ def test_aes128_cbc():
     #encdata = aes_encrypt(key, iv, data)
     encdata = unhexlify("4b0d438b4a1b972bd4ab81cd64674dcce4b0158090fbe616f455354284d53502")
 
-    ctx = rarfile.AES_CBC_Decrypt(key, iv)
+    ctx = AES_CBC_Decrypt(key, iv)
     assert ctx.decrypt(encdata) == data
 
 
-@pytest.mark.skipif(not rarfile._have_crypto, reason="No crypto")
+@pytest.mark.skipif(not have_crypto, reason="No crypto")
 def test_aes256_cbc():
     data = b"0123456789abcdef" * 2
     key = b"\x52" * 32
@@ -45,5 +43,5 @@ def test_aes256_cbc():
     #encdata = aes_encrypt(key, iv, data)
     encdata = unhexlify("24988f387592e4d95b6eaab013137a221f81b25aa7ecde0ef4f4d7a95f92c250")
 
-    ctx = rarfile.AES_CBC_Decrypt(key, iv)
+    ctx = AES_CBC_Decrypt(key, iv)
     assert ctx.decrypt(encdata) == data
