+++ test/test_format.py
@@ -237,35 +237,38 @@ def test_rar5_times():
 
 
 def test_oldvols():
-    assert rarfile._next_oldvol("archive") == "archive.r00"
-    assert rarfile._next_oldvol("archive.rar/foo") == "archive.rar/foo.r00"
-    assert rarfile._next_oldvol("archive.arr") == "archive.a00"
-    assert rarfile._next_oldvol("archive.brar") == "archive.b00"
-    assert rarfile._next_oldvol("qq00.part0.rar") == "qq00.part0.r00"
-    assert rarfile._next_oldvol("qq00.part0.r00") == "qq00.part0.r01"
-    assert rarfile._next_oldvol("qq00.part0.r29") == "qq00.part0.r30"
-    assert rarfile._next_oldvol("qq00.part0.r99") == "qq00.part0.s00"
+    from rarfile.format import _next_oldvol
+    assert _next_oldvol("archive") == "archive.r00"
+    assert _next_oldvol("archive.rar/foo") == "archive.rar/foo.r00"
+    assert _next_oldvol("archive.arr") == "archive.a00"
+    assert _next_oldvol("archive.brar") == "archive.b00"
+    assert _next_oldvol("qq00.part0.rar") == "qq00.part0.r00"
+    assert _next_oldvol("qq00.part0.r00") == "qq00.part0.r01"
+    assert _next_oldvol("qq00.part0.r29") == "qq00.part0.r30"
+    assert _next_oldvol("qq00.part0.r99") == "qq00.part0.s00"
 
 
 def test_newvols():
-    assert rarfile._next_newvol("qq00.part0.rar") == "qq00.part1.rar"
-    assert rarfile._next_newvol("qq00.part09.rar") == "qq00.part10.rar"
-    assert rarfile._next_newvol("qq00.part99.rar") == "qq00.part100.rar"
-    assert rarfile._next_newvol("part20") == "part21.rar"
-    assert rarfile._next_newvol("qq00.part3.exe") == "qq00.part4.rar"
-    assert rarfile._next_newvol("qq00.part5.sfx") == "qq00.part6.rar"
-    assert rarfile._next_newvol("qq00.part6.bin") == "qq00.part7.bin"
-    assert rarfile._next_newvol("99") == "100.rar"
-    assert rarfile._next_newvol("dir/99.rar") == "dir/100.rar"
+    from rarfile.format import _next_newvol
+    assert _next_newvol("qq00.part0.rar") == "qq00.part1.rar"
+    assert _next_newvol("qq00.part09.rar") == "qq00.part10.rar"
+    assert _next_newvol("qq00.part99.rar") == "qq00.part100.rar"
+    assert _next_newvol("part20") == "part21.rar"
+    assert _next_newvol("qq00.part3.exe") == "qq00.part4.rar"
+    assert _next_newvol("qq00.part5.sfx") == "qq00.part6.rar"
+    assert _next_newvol("qq00.part6.bin") == "qq00.part7.bin"
+    assert _next_newvol("99") == "100.rar"
+    assert _next_newvol("dir/99.rar") == "dir/100.rar"
     with pytest.raises(rarfile.BadRarName):
-        rarfile._next_newvol("qq00.part7.rar/foo")
+        _next_newvol("qq00.part7.rar/foo")
     with pytest.raises(rarfile.BadRarName):
-        rarfile._next_newvol("foo")
+        _next_newvol("foo")
 
 
 def test_newvols_err():
+    from rarfile.format import _next_newvol
     with pytest.raises(rarfile.BadRarName):
-        rarfile._next_newvol("xx.rar")
+        _next_newvol("xx.rar")
 
 
 @pytest.mark.parametrize("fn", ["test/files/rar3-versions.rar", "test/files/rar5-versions.rar"])
